pyproject.toml
[tool.poetry]
name = "ice-breaker"
version = "0.1.0"
description = ""
authors = ["Seyed Morteza Hosseini <seyedmortezahosseini@Seyeds-MacBook-Pro.local>"]
readme = "README.md"
packages = [{include = "ice_breaker"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.352"
openai = "^1.6.1"
googlesearch-python = "^1.2.3"
linkedin-api = {git = "https://github.com/tomquirk/linkedin-api.git"}
lxml = "^5.0.0"
pydantic = "^2.5.3"
flask = "^3.0.0"
jinja-partials = "^0.1.1"


[tool.poetry.group.test.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 79

[[tool.mypy.overrides]]
module = "langchain.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "flask.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydantic.*"
ignore_missing_imports = true
